feat: Introduce EventData and Participant interfaces for event management
- Added EventData interface for detailed calendar event information, including properties for participants and event logistics.
- Introduced Participant interface to manage event participants in forms.
- Created EventFormState interface to handle the state of the event addition form.
- Updated useEventAdd and useMainView hooks to utilize the new interfaces for improved type safety and clarity.
- Refactored EventDetailDialog and its hook to import the new EventData type, enhancing code organization.
